# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: CareerCopilot API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /resumes:
    get:
      operationId: listResumes
      tags: [resumes]
      summary: List user's resumes
      security: [{ clerkAuth: [] }]
      responses:
        "200":
          description: List of resumes
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Resume"
    post:
      operationId: createResume
      tags: [resumes]
      summary: Upload or create a resume
      security: [{ clerkAuth: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ResumeInput"
      responses:
        "201":
          description: Resume created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Resume"
  /resumes/{id}:
    get:
      operationId: getResume
      tags: [resumes]
      summary: Get a specific resume
      security: [{ clerkAuth: [] }]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Resume found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Resume"
        "404":
          description: Not found
    patch:
      operationId: updateResume
      tags: [resumes]
      summary: Update a resume
      security: [{ clerkAuth: [] }]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ResumeUpdate"
      responses:
        "200":
          description: Resume updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Resume"
    delete:
      operationId: deleteResume
      tags: [resumes]
      summary: Delete a resume
      security: [{ clerkAuth: [] }]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted

  # ── Applications ──────────────────────────────────────────────────────────
...
components:
  securitySchemes:
    clerkAuth:
      type: http
      scheme: bearer
  schemas:
...
    Resume:
      type: object
      required: [id, userId, title, content, createdAt, updatedAt]
      properties:
        id:
          type: integer
        userId:
          type: string
        title:
          type: string
        content:
          type: string
        isDefault:
          type: boolean
        createdAt:
          type: string
        updatedAt:
          type: string
    ResumeInput:
      type: object
      required: [title, content]
      properties:
        title:
          type: string
        content:
          type: string
        isDefault:
          type: boolean
    ResumeUpdate:
      type: object
      properties:
        title:
          type: string
        content:
          type: string
        isDefault:
          type: boolean

    # ── Application ───────────────────────────────────────────────────────
...