# lib/api-zod/src/generated/types/resumeUpdate.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * CareerCopilot API specification
 * OpenAPI spec version: 0.1.0
 */

export interface ResumeUpdate {
  title?: string;
  content?: string;
  isDefault?: boolean;
}